requirements.txt
streamlit
requests
cryptography
openai
python-dotenv